+++ src/ViewHoldingsRecord.test.js
@@ -18,9 +18,18 @@ const defaultProps = {
   id: 'id',
   goTo: jest.fn(),
   holdingsrecordid: 'holdingId',
-  referenceTables: { holdingsSources: [{ id: 'sourceId' }], locationsById: {} },
+  referenceTables: {
+    holdingsSources: [{ id: 'sourceId' }],
+    locationsById: {
+      inactiveLocation: { name: 'Location 1', isActive: false },
+    },
+  },
   resources: {
-    holdingsRecords: { records: [{ sourceId: 'sourceId' }] },
+    holdingsRecords: {
+      records: [
+        { sourceId: 'sourceId', temporaryLocationId: 'inactiveLocation', }
+      ],
+    },
     instances1: { records: [{ id: 'instanceId' }], hasLoaded: true },
     permanentLocation: { hasLoaded: true },
     temporaryLocation: { hasLoaded: true },
@@ -122,6 +131,15 @@ describe('ViewHoldingsRecord actions', () => {
       const link = document.querySelector('#holdings-list-bound-with-items a.itemHrid');
       expect(link)
         .toHaveAttribute('href', '/inventory/?qindex=hrid&segment=items&query=' + id);
+    });
+  });
+
+  it('should display "inactive" by an inactive temporary location', async () => {
+    renderViewHoldingsRecord();
+
+    await waitFor(() => {
+      const tempLocation = document.querySelector('*[data-test-id=temporary-location]').innerHTML;
+      expect(tempLocation).toContain('Inactive');
     });
   });
 });